refactor: Issue에서 사용하는 DTO클래스파일 이동
