+++ backend/src/main/java/com/team5/issue_tracker/issue/dto/response/IssueMilestonePageResponse.java
@@ -1,6 +1,8 @@
-package com.team5.issue_tracker.milestone.dto.response;
+package com.team5.issue_tracker.issue.dto.response;
 
 import java.util.List;
+
+import com.team5.issue_tracker.milestone.dto.response.MilestoneSummaryResponse;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
